+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 ### Added
 
 - request building area from Ohsome API and calculate fraction of area of interest covered by buildings ([#8](https://gitlab.heigit.org/climate-action/plugins/land-consumption/-/issues/8))
+- request parking lot area from Ohsome API, calculate land consumed by parking lots, adding one row to results table ([#9](https://gitlab.heigit.org/climate-action/plugins/land-consumption/-/issues/9))
 
 ## [Dummy](https://gitlab.heigit.org/climate-action/plugins/land-consumption/-/releases/Dummy)
 - output dummy table output with the percentage of an area of interest consumed by different land uses